Drop NaCl-specific e_flags bits and EI_OSABI setting
The variants that the e_flags bits refer to are obsolete.
Despite its name, EI_OSABI has an OS-independent meaning in
upstream binutils and glibc ld.so, so we cannot use it.
We don't really need these distinguishing marks in a NaCl ELF file.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=564
TEST=trybots

R=sehr@google.com

Review URL: http://codereview.chromium.org/7453038

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@6280 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2