Extend IRT with nacl_irt_resource_open interface.
We need to solve the problem that we need to provide pointers to PPP_Initialize and PPP_GetInterface in the program before manifest file loading is working and so this program can be loaded.

In this CL, we split PpapiPluginMain in two parts so that ld.so can call the first part, load program, and then call the second part to provide pointers to these functions. This way is preferable since it make code in ld.so simplier.

New IRT interface provides a single function that returns a file descriptor by file name. The name service and manifest service channels are initialized at first use of this function. If an error happens during initialization, the function returns EIO.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2127
TEST= run_irt_manifest_file_chrome_browser_test
Review URL: http://codereview.chromium.org/7605029

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@6453 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2