Replace newlib's malloc with one of our own
Build our own copy of dlmalloc 2.8.5 as the allocator for the newlib-based
C library (i.e. put it in libnacl).  We are already using a different
incarnation of this same source for the IRT's private allocator, and it
works well.  The malloc implementation in newlib is based on a much older
version of dlmalloc, and it's also using pthread_mutex_lock for its
locking, while here we use simple spin locks.

Note: This must be committed along with a toolchain DEPS update that
brings in a newlib build that excludes its malloc.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2271
TEST= trybots

R=pasko@google.com,jvoung@google.com

Review URL: http://codereview.chromium.org/7991006

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@6759 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2