Print diagnostic message via NaClLog, not stdout
Sometimes we want to talk with a nacl module via stdin/stdout,
another option is just to fprintf(stderr).

This helps communicating with a chrooted untrusted process in
NaCl sandbox via stdin/stdout. The outside process that we pass
messages to is in a prototype stage now.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/8037027

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@6790 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2