Fix DEPS for 64 bit crt_platform on Windows
With no GYP based IRT test in NaCl and nacl_integration disabled,
we missed a failure where the IRT was not getting built for
Windows 64 bit.

Normally the compile step would have caught this, however for
Windows 64 bit GYP generation with multiple actions we incorrectly
use multiple calls without a '&&'.  So the success of the second
action hides the failure of the first.

This is being address in a seperate CL.

To prevent this type of error in the future, this CL not only
fixes the missing OS='win' check which enables 64 bit, but also
adds a test that retures sel_ldrXXX, irt_core_XXX.nexe, and
hello_world_XXX.nexe to successfully build and run as part of
the compile step.


TEST= try
BUG= http://code.google.com/p/chromium/issues/detail?id=96782
Review URL: http://codereview.chromium.org/8450019

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7105 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2