Fix gsutil invocations on coverage bots
The gsutil_cp_dir.py script has changed so that the GSUTIL is treated
as one word regardless of spaces inside it.  This broke the uses in our
coverage bots where we pass a flag inside that variable.
We don't really need to specify -h Cache-Control:no-cache, so drop it.

BUG= none
TEST= none

R=bradnelson@google.com

Review URL: http://codereview.chromium.org/8463030

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7143 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2