remove javascript console log message (and test thread)
with this CL, the ability to display strings on the javascript console
becomes untested / unexercised.  the next step is to plumb through the
ability to send untrusted stdout/stderr to the javascript console.
i'm in favor of allowing NACL_EXE_STDOUT and NACL_EXE_STDERR take on
special values to specify the javascript console (as opposed to a
file), since this will make printf-debugging the nexe easier, since
it'll be on-screen and not require the --no-sandbox flag (though the
output would then become transient).

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2115
TEST= None
Review URL: http://codereview.chromium.org/8488009

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7253 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2