x86-64 Windows: Make stack unwinding test stricter still
Add a check that the return address has unwind info.  When I committed
the initial version of this test I had not found the right API for
doing this.

SymFunctionTableAccess64() ought to work for this purpose too, though
it returns a different address -- presumably a different copy of the
unwind info, since this function works across processes and is
provided by dbghelp.lib rather than the runtime library.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2414
TEST=run_windows_crash_in_syscall_test

Review URL: http://codereview.chromium.org/8769015

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7374 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2