Revert r7374: x86-64 Windows: Make stack unwinding test stricter still
This change failed to compile on the 64-bit Windows Buildbots, which
apparently differ from the trybots in the version of the Windows
headers that are installed.  UNWIND_HISTORY_TABLE and
RtlLookupFunctionEntry() are not defined.  The errors indicate that
the header being used is:
C:\Program Files\Microsoft SDKs\Windows\v6.0A\include\winnt.h

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2414
TEST=none

Review URL: http://codereview.chromium.org/8825006

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7375 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2