Add platform to standard buildbot step names in toolchain bot case
The toolchain bots run the standard buildbot set twice, once as x86-32 and
once as x86-64.  This means the same step names appear twice in the output
and thus the waterfall, which is confusing.  So add the platform name to
the step names when "inside" the toolchain bot.

BUG= none
TEST= trybots

R=bradnelson@google.com,ncbray@google.com

Review URL: http://codereview.chromium.org/8949002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7461 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2