Change the x86-32 validator to work in offsets rather than physical addresses.
Previously the validator computed jump addresses in the memory region the code
was being mapped into rather than relative to the start of the code segment.
This obscured the fact that the validation algorithm did not actually depend
on the base address, as long as it was correctly aligned.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2515
TEST= ./scons platform=x86-32 validator_tests

Review URL: http://codereview.chromium.org/9190032

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7585 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2