libnacl changes to enable clock_get{res,time}.  moves timespec defn.
This CL adds clock_getres and clock_gettime functions to the nacl
library, used to invoke the corresponding functions in the IRT.  It
also moves the definition for the timespec structure to
machine/_types.h to be defined at the same place as time_t, to work
with the newlib time.h which includes machine/_types.h but not
sys/time.h.  This latter change is a precondition for a newlib
toolchain patch to expose the correct values for CLOCK_MONOTONIC etc
via sys/features.h and time.h changes.

the clock_getres and clock_gettime functions cannot be tested until
after the corresponding toolchain change has been committed, so the
tests are in a separate CL.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2477
TEST= existing tests.
Review URL: https://chromiumcodereview.appspot.com/9121038

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7691 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2