Exception handling: Ensure exception_stack and exception_flag are uint32_t
nacl_reg_t is 64 bits on x86-64.  Storing exception_stack as 64-bit
carries a greater risk of accidentally writing a stack outside of
untrusted address space, so change it to uint32_t.

This is in preparation for exception handling on x86-64 Windows, where
we will want to read exception_stack with ReadProcessMemory().

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2536
TEST=exception_tests and others

Review URL: https://chromiumcodereview.appspot.com/9443032

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7880 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2