Exception handling: Move exception_stack/exception_flag to NaClAppThread
This is simpler since the fields aren't at all arch-specific.  It
saves a little space since the fields won't be duplicated in the
'user' and 'sys' members of NaClAppThread.

exception_stack/exception_flag were originally put into
NaClThreadContext on the grounds that because NaClThreadContext is
commented with integer field offsets, this might be a stable
interface.  But really, the offsets are commented for internal use by
service_runtime's assembly code, and are not meant to be a stable
interface.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2536
TEST=exception_tests

Review URL: https://chromiumcodereview.appspot.com/9516012

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@7909 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2