Fix linking of sel_main_chrome.c (broken by r8019)
Fix linking on Unix: Make sure that NaClSetCreateMemoryObjectFunc() is
defined outside of "namespace nacl" so that is compiled as 'extern
"C"'.  Move the global variables to the top of the file.  Move
NaClDuplicateNaClHandle() to keep the 'extern "C"' functions together.

Fix linking on Windows: Add an empty NaClSetCreateMemoryObjectFunc()
for Windows.

Fix the definition of NaClChromeMainArgsCreate() to use the correct
name!

Add a test to cover linking of sel_main_chrome.c.

r8019 broke Chromium integration: I had tested building Chromium
locally, on a Mac, but I only tested an earlier version of the change.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2651
TEST=trybots

Review URL: https://chromiumcodereview.appspot.com/9703033

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8027 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2