Enable support for _SC_PAGESIZE at the sysconf syscall level.
This enables newlib sysconf() to query and obtain the "page size" of
64kB.  The test is conditional on __GLIBC__ because the _SC_PAGESIZE
macro is not available in the current newlib toolchain -- after the
next TC roll that code can be removed.

glibc believes that it owns the configuration names in
bits/confnames.h and the glibc port already handles _SC_PAGESIZE.  we
could make newlib's sysconf wrapper just intercept and shunt the
syscall when the confid is _SC_PAGESIZE as well, but the confname
namespace used in newlib comes from the exported ABI definition file,
and it's trivial to support _SC_PAGESIZE at the syscall level.  we
could still implement the syscall shunting for the toolchain-side
change, but i don't think it's necessary (this is infrequently used)
and don't want to increase the number of places where the magic 64k
number is defined.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2632
TEST= run_sysconf_test
Review URL: https://chromiumcodereview.appspot.com/9969125

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8184 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2